feat: cuda tiled multiplication working build

+++ include/matrix_transform/matrix_types.hpp
@@ -1,3 +1,3 @@
 #include <Eigen/Dense>
 
-using Matrix = Eigen::MatrixXd;
+using Matrix = Eigen::MatrixXf;
